Add content to recipe page
Create a directory called "image" to store the images for the different
recipes.

In the HTML file for the first recipe, add a free image of the finished
recipe, and h2 headings for sections like "Description," "Ingredients,"
and "Steps." Use p elements for the first section, and ul and ol elements
for the last two, respectively.

This will provide the user with complete recipe information.